REF: move funcs to graphics, network
